requirements.txt
Flask
Flask-GraphQL
Flask-Cors
python-dotenv
requests
keycloak
stripe
graphql-core
tinydb
